✂️🎃 Added pagination with sqlite and media Recognition

+++ plugins/dl/dl.py
@@ -1,11 +1,11 @@
-
 from utils.typing import *
+import utils.menus
+import traceback 
+DOWNLOAD_LOCATION = "./DOWNLOADS"
 
 
-#getd = DownL("https://bfas237blog.com/wp-content/uploads/2018/12/bfas237blog-transpa.png")
-# Fix extensions
 @Client.on_message(Filters.regex("dl_"))
-def my_handler(bot, m):
+def my_handler(bot, m, *args, **kwargs):
     chat_id = m.chat.id
     user = m.from_user.id
     h = m.text
@@ -14,62 +14,67 @@ def my_handler(bot, m):
     else:
       g = h[4:]
     tnews = fileid(g) 
+    file =  tnews
     #m.reply(tnews)
     apk_string = "{}".format("apks")
-    try:
-      if tnews:
-        bot.send_chat_action(chat_id,'UPLOAD_DOCUMENT')
-        time.sleep(1)
-        bot.send_document(m.chat.id, tnews, thumb="bfas237blog-transpa.png", caption="Powered with ❤️ - @Bfas2327Bots")
-      else:
+    if tnews:
+        try:
+           
+          bot.send_chat_action(chat_id,'UPLOAD_DOCUMENT')
+          time.sleep(1)
+          bot.send_document(m.chat.id, tnews, caption="Powered with ❤️ - @Bfas2327Bots")
+        except FileIdInvalid:
+          okok(bot, m,  g)
+    else:
+        
+        bot.send_chat_action(chat_id,'TYPING')
+        time.sleep(1.5)
+        m.reply("😔 `404` **Invalid FILE_ID**.\n\n Send /files to see your saved files")
+    
+       
+
+
+def okok(bot, m,  fid):
+    chat_id = m.chat.id
+    user = m.from_user.id
+    h = m.text
+    if "/start" in h:
+      g = h[10:]
+    else:
+      g = h[4:]
+    tnews = fileid(fid) 
+    file =  tnews
+    #m.reply(tnews)
+    apk_string = "{}".format("apks")
+    if tnews:
+        try:
+           
+          bot.send_chat_action(chat_id,'UPLOAD_VIDEO')
+          
+          time.sleep(1)
+          bot.send_video(m.chat.id, tnews, caption="Powered with ❤️ - @Bfas2327Bots")
+          pass
+        except FileIdInvalid:
+          bot.send_audio(m.chat.id, tnews, caption="Powered with ❤️ - @Bfas2327Bots")
+          traceback.print_exc()   
+    else:
         
         bot.send_chat_action(chat_id,'TYPING')
         time.sleep(1.5)
         m.reply("😔 `404` **Invalid FILE_ID**.\n\n Send /files to see your saved files")
  
-    except Exception as e:
-        m.reply(str(e))
+       
+
 
        
-@Client.on_message(Filters.command("files"))
-def sendServerStartedMessage(bot, m):
-    link = " ".join(m.command[1:])
-    chat_id = m.chat.id
-    bot.send_chat_action(chat_id,'TYPING')
-    # Connecting to the SQL database
-    conn = sqlite3.connect('inshorts.db')
-    c = conn.cursor()
-    con = conn.cursor()
 
-    dd = m.from_user.id
-    chat_id = str(chat_id)
-    c.execute("SELECT ID, Fname, DownloadId, Size FROM files WHERE User = ?", (chat_id, ))
-    con.execute("SELECT COUNT (*) FROM files")
-    rowcount = con.fetchone()[0]
-    rows = c.fetchall() 
-    conn.close()
-    try: 
-      if len(rows) > 0: 
-        items = ""
-        lens = len(rows)
-        for row in rows:
-            items +=  (
-              "<code>#{}</code> " 
-              " <b>{}</b>"     
-              "\n<a href='https://telegram.me/megabookstorebot?start=dl_{}'>Click to download</a>\n"       
-              "<i>{}</i>\n"  
-              "------\n" 
-              "".format(str(row[0]), row[1], row[2], pretty_size(int(row[3]))))
-        username = m.from_user.username
-        m.reply("📄 <b>{}</b>'s files Library: <b>{}</b> \n\n {}".format(username, rowcount, items), parse_mode="html")
-      else:
-        m.reply("No items in your list")
- 
-    except Exception as e:
-        m.reply(str(e))
       
-      
-      
+     
+
+
+def ss(client, callback_query):
+    client.answer_callback_query(callback_query.id, "it works hhhh! {}".format(callback_query.from_user.first_name), show_alert=True)
+  
 
 @Client.on_message(Filters.command("get"))
 def dl(client, message, **current):
@@ -89,7 +94,6 @@ def dl(client, message, **current):
             if fname:
                 fnames, ext = fname 
                 required_file_name = fnames+ext
-                message.reply(required_file_name)
             else:
                 message.reply("😭 Invalid file name... retrying")
                 time.sleep(5)
@@ -98,7 +102,7 @@ def dl(client, message, **current):
                 required_file_name = os.path.basename(fnames) 
             sent = message.reply("**Checking:** `{}`\n\n if it exist on my server..".format(required_file_name), quote=True, reply_to_message_id=message.message_id)
             time.sleep(2)
-            lr = checkUserLastNews(message.chat.id)
+            lr = checkUserLastNews(message.from_user.id)
             tr = checkTodayFirstNewsID()
             tnews = "No files so let me download it for you"
             ttfiles = "Could not send requested file"
@@ -113,17 +117,25 @@ def dl(client, message, **current):
             if(tr != 0):
               tnews, size = sfileid(url) 
             if(tnews != 0):
+              time.sleep(2)
               client.send_document(message.chat.id, tnews, caption="Oh! i had this alread so it was faster")
               sent.delete()
             elif(size == 0):
               rd = 1
               sent.delete()
-              sent = message.reply("**Downloading...:**", quote=True, reply_to_message_id=message.message_id)
+              start = time.time()
+              sent = message.reply("**Downloading...**", quote=True, reply_to_message_id=message.message_id)
+              time.sleep(3)
               downl = DownLoadFile(url, required_file_name, client, sent, message.chat.id)
+              time.sleep(3)
               if downl:
+                if (downl != 0):
                   sent.edit("Done ✅.. Now uploading..")
-                  chat_id = message.chat.id
-                  file = client.send_document(message.chat.id, required_file_name, progress = prog, progress_args = (sent, message.chat.id, required_file_name))  
+                  time.sleep(3)
+                  chat_id = message.from_user.id
+                  required_file_name = download_path+"/"+required_file_name
+                  file = client.send_document(message.chat.id, required_file_name, progress = DFromUToTelegramProgress, progress_args = (sent, message.chat.id, start))  
+                  
                   os.remove(required_file_name)
                   download_id = generate_uuid()
                   file_size = file.document.file_size
@@ -140,9 +152,11 @@ def dl(client, message, **current):
                   LastReadNewsID = checkUserLastNews(chat_id)
                   TodayFirstNewsID = checkTodayFirstNewsID()
                   news = "No news"
+                  
+                  
                   tfiles = None 
                   if(TodayFirstNewsID == 0):
-                    news = "No news for today."
+                    news = "Oh i encountered an error while saving."
                   elif(LastReadNewsID < TodayFirstNewsID):
                     LastReadNewsID = TodayFirstNewsID
                   if(TodayFirstNewsID != 0):
@@ -157,10 +171,11 @@ def dl(client, message, **current):
                 er = "An error occured while downloading...."
                 sent.edit(er)
                 logger.info(er)
+                
         else:
           err = "`Your link doesn't look like a downloadable link...... Kindly try again`"
           message.reply(str(err), quote=True)
           logger.debug(str(err))
         
     except:
-      traceback.print_exc()
+      traceback.print_exc()
